CacheMemory: Moving template implementation to .hpp
